The MetaMask struggle continues (#107)
MetaMask has two presentations, one is an overlay that is always on top of the browser and the other is a popup which can disappear behind the browser if the user is not careful. The popup is forced open by eht_requestAccounts or wallet_requestPermissions. The former is a nightmare because calling it twice with the popup behind causes things to lock up. The user can only fix it by using the overlay. The later works better but still requires teh user to use the overlay for adjustments.